+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/ProcessesResources.scala
@@ -106,7 +106,7 @@ class ProcessesResources(
               }
             }
           }
-        } ~ path("processes" / "statusDefinitions") {
+        } ~ path("processes" / "stateDefinitions") {
           get {
             complete {
               fetchAllStateDefinitions