+++ docs/Changelog.md
@@ -25,6 +25,10 @@
   thanks to that status and versions panel are refreshed with "DEPLOYED" state in the same time
 * [#3992](https://github.com/TouK/nussknacker/pull/3992) Fix for compiling scenarios containing filter node with only 'false' edge
 * [#4127](https://github.com/TouK/nussknacker/pull/4127) ResourceLoader and bumps commons io 2.4 -> to 2.6
+* [#4066](https://github.com/TouK/nussknacker/pull/4066) 
+  * Add state definitions to `ProcessStateDefinitionManager`
+  * Add `ProcessResources` endpoint that returns state definitions with default state properties (such as displayable name, icon and description), 
+    to allow filtering by state in UI.
 
 1.8.1 (28 Feb 2023)
 ------------------------